[refactor] user 서비스 디렉토리 구조 수정
