+++ services/user/src/main/java/com/ticketPing/user/application/dto/UserResponse.java
@@ -1,5 +1,5 @@
 package com.ticketPing.user.application.dto;
-import com.ticketPing.user.domain.entity.User;
+import com.ticketPing.user.domain.model.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
