+++ services/user/src/main/java/com/ticketPing/user/infrastructure/repository/UserJpaRepository.java
@@ -1,6 +1,6 @@
 package com.ticketPing.user.infrastructure.repository;
 
-import com.ticketPing.user.domain.entity.User;
+import com.ticketPing.user.domain.model.entity.User;
 import com.ticketPing.user.domain.repository.UserRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
