+++ services/user/src/test/java/com/ticketPing/user/CustomerApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class UserApplicationTests {
+class CustomerApplicationTests {
 
 	@Test
 	void contextLoads() {
